+++ src/s3m/mod.rs
@@ -5,4 +5,4 @@ mod upload;
 pub use self::upload::{multipart_upload, upload};
 
 mod streams;
-pub use self::streams::{Stream, DB_PARTS, DB_UPLOADED};
+pub use self::streams::Stream;