+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Using [sled](http://sled.rs/) with [serde_cbor](https://crates.io/crates/serde_cbor) to keep track of the uploaded files, thanks @D1plo1d
 * Multipart uploads can be resumed and is the default behavior.
 * The file checksum and the s3 credentials are used to keep track of the uploads this also prevent uploading the same file multiple times to the same location.
+* Added option `-r` to cleanup `~/.s3m/streams` directory.
 
 ## 0.2.0
 * Implemented lifetimes  🌱